arm build (#61)
* First pass at porting to ARM:
 - SHMEM and Mercury are x86-only features
 - Unit tests pass

* Timestamp Demo setup scripts:
 - Added configuration resources directory
 - Renamed serial udev rules file
 - Renamed script for updating serial udev rules

* Time Demo: Updated serial device names

* Time Demo: Added configuration files and a script for configuring network resources

* PNT Demo: Fixed destination parameter for the 5.yield version

* Switch (ssize_t) cast to (size_t) cast

+++ demos/channel_demo/src/reader.c
@@ -103,7 +103,7 @@ static int reader_run(reader_t *reader) {
             }
         }
 
-        if (rd_len != exp_len) {
+        if (((size_t) rd_len) != exp_len) {
             bin_save("exp", out_dir, exp_buf, exp_len);
             bin_save("got", out_dir, rd_buf, rd_len);
             log_msg(ERROR, "Expecting %u bytes, got %u. Binary files saved",
